backend/pyproject.toml
[project]
name = "osint-framework-backend"
version = "0.1.0"
description = "Django Backend for OSINT Investigation Framework"
requires-python = ">=3.10"
dependencies = [
  "django>=4.2.7",
  "djangorestframework>=3.14.0",
  "python-dotenv",
  "psycopg2-binary",
]
[project.optional-dependencies]
dev = ["pytest", "pytest-django", "coverage", "mypy"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "settings"
python_files = "test_*.py"
addopts = "-q --reuse-db"
